refactor: :truck: move step components out of subcomponents folder
