+++ src/renderer/components/ProjectDetailsInput/style.ts
