+++ src/renderer/components/ProjectDetailsInput/validationSchema.ts
